bug fix - sometimes heading and speed are not sent in the webservice xml payload.
git-svn-id: http://svn.code.sf.net/p/nextbusapi/code/trunk@101 56530abd-8947-4513-bbf7-f0464c63a0cc

+++ adapter/src/main/java/net/sf/nextbus/publicxmlfeed/impl/DomainFactory.java
@@ -253,7 +253,7 @@ public class DomainFactory {
             long lastTime = System.currentTimeMillis() - v.getSecsSinceReport() * 1000;
 
             VehicleLocation vehLcn = new VehicleLocation(route, v.getId(), v.getDirTag(),
-                    v.isPredictable().booleanValue(), lastPosition, lastTime, v.getSpeedKmHr().doubleValue(), v.getHeading().doubleValue(), cpyRt);
+                    v.isPredictable().booleanValue(), lastPosition, lastTime, v.getSpeedKmHr(), v.getHeading(), cpyRt);
             returnValue.add(vehLcn);
         }
         return returnValue;
